% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_data_analysis.R
\name{plot_cumulative_time_in_space}
\alias{plot_cumulative_time_in_space}
\title{Plot Cumulative Time in Space Over the Years}
\usage{
plot_cumulative_time_in_space(tdf, graph_file)
}
\arguments{
\item{tdf}{A dataframe containing a "duration" column in "HH:MM" format and a "date" column.}

\item{graph_file}{A character string specifying the path to save the graph.}
}
\description{
This function plots the cumulative time spent in space over the years based on
the data in the dataframe. The cumulative time is calculated by converting the
"duration" column into hours, then computing the cumulative sum of the duration.
The plot is saved as a PNG file at the specified location.
}
