% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_data_analysis.R
\name{read_json_to_dataframe}
\alias{read_json_to_dataframe}
\title{Read and Clean EVA Data from JSON}
\usage{
read_json_to_dataframe(input_file)
}
\arguments{
\item{input_file}{A character string specifying the path to the input JSON file.}
}
\value{
A cleaned and sorted data frame containing the EVA data.
}
\description{
This function reads EVA data from a JSON file, cleans it by converting
the 'eva' column to numeric, converting data from text to date format,
the data by the 'date' column.
}
