% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_data_analysis.R
\name{text_to_duration}
\alias{text_to_duration}
\title{Convert Duration from HH:MM Format to Hours}
\usage{
text_to_duration(duration)
}
\arguments{
\item{duration}{A character string representing the duration in "HH:MM" format.}
}
\value{
A numeric value representing the duration in hours.
}
\description{
This function converts a duration in "HH:MM" format (as a character string)
into the total duration in hours (as a numeric value).
}
\details{
When applied to a vector, it will only process and return the first element
so this function must be applied to a data frame rowwise.
}
\examples{
text_to_duration("03:45")  # Returns 3.75 hours
text_to_duration("12:30")  # Returns 12.5 hours
}
